backend/requirements.txt
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
flask==3.1.1
serverless-wsgi==3.1.0
flask-sqlalchemy==3.1.1
Flask-JWT-Extended~=4.7.1
flask-cors==6.0.1
google-auth==2.40.3
